conexão sql server

+++ src/main/java/view/inc/model/ProcessoIlicito.java
@@ -2,9 +2,11 @@ package view.inc.model;
 
 import com.github.britooo.looca.api.core.Looca;
 import view.inc.dao.ProcessoDao;
+import view.inc.dao.ProcessoSQLDao;
 
 import java.io.IOException;
 import java.sql.Date;
+import java.sql.SQLException;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -17,17 +19,18 @@ public class ProcessoIlicito{
 
     private List<ProcessoIlicito> processosIlicitos = new ArrayList<>();
     private ProcessoDao processoDao = new ProcessoDao();
+    private ProcessoSQLDao processoSQLDao = new ProcessoSQLDao();
 
 
-    public ProcessoIlicito(){
+    public ProcessoIlicito() throws SQLException, ClassNotFoundException, InstantiationException, IllegalAccessException {
     }
 
-    public ProcessoIlicito(Integer idProcesso, int fkProcesso, String nomeProcesso) {
+    public ProcessoIlicito(Integer idProcesso, int fkProcesso, String nomeProcesso) throws SQLException, ClassNotFoundException, InstantiationException, IllegalAccessException {
         this.idProcessoIlicito = idProcesso;
         this.nomeProcesso = nomeProcesso;
         this.fkProcesso = fkProcesso;
     }
-    public ProcessoIlicito(int fkProcesso, String nomeProcesso) {
+    public ProcessoIlicito(int fkProcesso, String nomeProcesso) throws SQLException, ClassNotFoundException, InstantiationException, IllegalAccessException {
         this.nomeProcesso = nomeProcesso;
         this.fkProcesso = fkProcesso;
     }
@@ -41,7 +44,7 @@ public class ProcessoIlicito{
         return this.processosIlicitos;
     }
 
-    public void checkProcessosIlicitos(Computador computador) throws IOException {
+    public void checkProcessosIlicitos(Computador computador) throws IOException, SQLException, ClassNotFoundException, InstantiationException, IllegalAccessException {
         Janela janela = new Janela();
         List<com.github.britooo.looca.api.group.janelas.Janela> janelas = janela.getAllJanelas();
         List<ProcessoIlicito> processoIlicitos = processoDao.selectAllProcessosIlicitos(computador.getIdComputador());
@@ -54,6 +57,28 @@ public class ProcessoIlicito{
                     int pid = janela1.getPid().intValue();
                     //adicionar aqui o registro de processo ilicito com data e hora da ocorrencia
                     processoDao.insertRegistroIlicito(nomeProcesso, computador.getIdComputador());
+                    //aqui pede pra matar aquele processo
+                    matarProc.killProcess(pid);
+                }
+            }
+        }
+        System.out.println("Matei todos os processos ilícitos");
+    }
+
+    public void checkProcIlicitosSQL(Computador computador) throws IOException, SQLException, ClassNotFoundException, InstantiationException, IllegalAccessException {
+        Janela janela = new Janela();
+        List<com.github.britooo.looca.api.group.janelas.Janela> janelas = janela.getAllJanelas();
+        List<ProcessoIlicito> processoIlicitos = processoSQLDao.selectAllProcessosIlicitos(computador.getIdComputador());
+        ProcessoComputador matarProc = new ProcessoComputador();
+        for (com.github.britooo.looca.api.group.janelas.Janela janela1 : janelas) {
+            for (int i = 0; i < processoIlicitos.size(); i++) {
+                String nomeProcesso = janela1.getTitulo().toLowerCase();
+                String nomeIlicito = processoIlicitos.get(i).getNomeProcesso().toLowerCase();
+                if(nomeProcesso.contains(nomeIlicito)){
+                    int pid = janela1.getPid().intValue();
+                    //adicionar aqui o registro de processo ilicito com data e hora da ocorrencia
+                    processoDao.insertRegistroIlicito(nomeProcesso, computador.getIdComputador());
+                    processoSQLDao.insertRegistroIlicito(nomeProcesso, computador.getIdComputador());
                     //aqui pede pra matar aquele processo
                     matarProc.killProcess(pid);
                 }
